[mod/#254] 채널 생성 함수 NazapFirebaseMessaging으로 이동
